requirements.txt
pyserial==3.5
setuptools==80.9.0
wheel==0.45.1
